+++ lib/tiled/model/tiled_item_tile_set.dart
@@ -7,6 +7,9 @@ class TiledItemTileSet {
   final List<CollisionArea>? collisions;
   final String? type;
   final Map<String, dynamic>? properties;
+  final double angle;
+  final bool isFlipVertical;
+  final bool isFlipHorizontal;
 
   TiledItemTileSet({
     this.sprite,
@@ -14,5 +17,8 @@ class TiledItemTileSet {
     this.animation,
     this.type,
     this.properties,
+    this.angle = 0,
+    this.isFlipVertical = false,
+    this.isFlipHorizontal = false,
   });
 }